backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.10
aiosignal==1.4.0
aiosqlite==0.20.0
alembic==1.14.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.3
APScheduler==3.11.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.0
asyncpg==0.30.0
attrs==23.1.0
axios==0.4.0
backcall==0.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
blis==0.7.11
boto3==1.41.2
botocore==1.41.2
bs4==0.0.2
cachetools==6.2.2
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.4
confection==0.1.5
cryptography==46.0.3
cymem==2.0.13
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
Django==4.2.15
dnspython==2.8.0
docopt==0.6.2
ecdsa==0.19.1
email_validator==2.2.0
executing==2.0.0
fake-useragent==2.2.0
fastapi==0.115.0
fastjsonschema==2.18.1
filelock==3.15.4
firecrawl-py==4.12.0
Flask==3.0.0
Flask-Cors==4.0.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.37
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.43.0
google-auth-httplib2==0.3.0
google-genai==1.4.0
googleapis-common-protos==1.72.0
googlemaps==4.10.0
greenlet==3.3.0
grip==4.6.2
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.17.3
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
ipykernel==6.25.2
ipython==8.16.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.12.0
jmespath==1.0.1
joblib==1.3.1
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-server-mathjax==0.2.6
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab_git==0.43.0
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.23.0
limits==5.6.0
lxml==6.0.2
Mako==1.3.10
mangum==0.20.0
marisa-trie==1.3.1
Markdown==3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.15
nbclient==0.8.0
nbconvert==7.9.2
nbdime==3.2.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.5
nltk==3.8.1
node==1.2.2
npm==0.1.1
numpy==1.26.1
odict==1.9.0
openai==2.16.0
optional-django==0.1.0
overrides==7.4.0
packaging==25.0
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
path-and-address==2.0.1
pdf2image==1.17.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
pipenv==2024.0.1
platformdirs==3.11.0
playwright==1.57.0
pluggy==1.6.0
plumber==1.7
polyline==2.0.4
postgrest==2.27.1
preshed==3.0.12
prometheus-client==0.17.1
prompt-toolkit==3.0.39
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.12.5
pydantic-settings==2.6.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pymongo==4.6.0
pyparsing==3.3.1
PyPDF2==3.0.1
pyroaring==1.0.3
pytesseract==0.3.10
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-docx==1.0.1
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-Levenshtein==0.23.0
python-multipart==0.0.20
python-telegram-bot==21.10
pythonanywhere==0.14.1
pythonanywhere-core==0.1.10
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
RapidFuzz==3.14.3
realtime==2.27.1
referencing==0.30.2
regex==2025.11.3
reportlab==4.0.7
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
-e git+https://github.com/Tjindl/AxiomForge.git@1aed38778dbd2769f8a898b208d7eb7b844ad23b#egg=rl_math&subdirectory=rl_math
rpds-py==0.10.4
rsa==4.9.1
s3transfer==0.15.0
safetensors==0.7.0
schedule==1.2.2
schema==0.7.7
scikit-learn==1.3.2
scipy==1.11.1
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
slowapi==0.1.9
smart-open==6.4.0
smmap==5.0.1
snakesay==0.10.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
sqlparse==0.5.1
srsly==2.5.2
stack-data==0.6.3
starlette==0.38.6
storage3==2.27.1
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.1
supabase-auth==2.27.1
supabase-functions==2.27.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.17.1
thinc==8.2.5
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.14.1
torch==2.9.1
tornado==6.3.3
tqdm==4.67.2
traitlets==5.11.2
transformers==4.35.0
tree-sitter==0.25.2
tree-sitter-cpp==0.23.4
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-rust==0.24.0
typer==0.9.4
types-python-dateutil==2.8.19.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.32.0
uvloop==0.22.1
virtualenv==20.26.3
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.8
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==14.2
Werkzeug==3.0.1
wrapt==2.0.1
yarl==1.22.0
zenrows==1.4.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==7.0.1
zope.lifecycleevent==5.0
zope.proxy==5.3
